+++ src/MultiplicationTables.java
@@ -3,17 +3,17 @@ import java.util.Scanner;
 public class MultiplicationTables {
     public static void main(String[] args) {
         System.out.println("Please input a number");
-        int inputNum = MultiplicationTables.scanner();
+        double inputNum = MultiplicationTables.scanner();
         MultiplicationTables.loop(inputNum);
     }
-    public static void loop(int inputNum) {
+    public static void loop(double inputNum) {
         System.out.println("User input the number "+inputNum);
         for (int i = 1; i <= 10; i++) {
             System.out.println(inputNum+" * "+i+" = "+(inputNum * i));
         }
     }
-    public static int scanner() {
+    public static double scanner() {
         Scanner scan = new Scanner(System.in);
-        return scan.nextInt();
+        return scan.nextDouble();
     }
 }
