+++ src/StopAtFive.java
@@ -4,14 +4,14 @@ public class StopAtFive {
     public static void main(String[] args) {
         while (true) {
             System.out.println("Give a number:");
-            int inputNum = StopAtFive.scanner();
-            if (Integer.valueOf(inputNum) == 5) {
+            double inputNum = StopAtFive.scanner();
+            if (inputNum == 5) {
                 break;
             }
         }
     }
-    public static int scanner() {
+    public static double scanner() {
         Scanner scan = new Scanner(System.in);
-        return scan.nextInt();
+        return scan.nextDouble();
     }
 }
